backend/requirements.txt
replicate
python-dotenv
groq
pyaudio 
SpeechRecognition
flask 
flask-cors
assemblyai
boto3